requirements-dev.txt
black
pre-commit
pytest
pytest-cov
pytest-django
